A multitude of bug fixes

+++ build.sh
@@ -4,6 +4,7 @@ PAK=$(cat DEBIAN/control | grep 'Package: ' | sed 's/Package: //g')
 ARCH=$(cat DEBIAN/control | grep 'Architecture: '| sed 's/Architecture: //g')
 FOLDER="$PAK\_$VERSION\_$ARCH"
 FOLDER=$(echo "$FOLDER" | sed 's/\\//g')
+OPTIONS="$1"
 mkdir ../"$FOLDER"
 ##############################################################
 #							     #
@@ -14,40 +15,48 @@ mkdir ../"$FOLDER"
 ##############################################################
 
 # Instead of compiling, we are building a tar.xz archive of the latest kernel package
-cd usr/share/system-installer/modules
-echo -e "\t###\tDOWNLOADING\t###\t"
-rsync -vr rsync://apt.draugeros.org/aptsync/pool/main/l/linux-upstream kernel
-rsync -vr rsync://apt.draugeros.org/aptsync/pool/main/l/linux-meta kernel
-echo -e "\t###\tDELETING CRUFT\t###\t"
-list=$(ls kernel)
-for each in $list; do
-	remove=$(ls kernel/$each | grep -v 'amd64.deb$')
-	for each2 in $remove; do
-		rm -rfv kernel/$each/$each2
+# Don't make the archive if --pool passed
+if [ "$OPTIONS" != "--pool" ]; then
+	cd usr/share/system-installer
+	echo -e "\t###\tDOWNLOADING\t###\t"
+	rsync -vr rsync://apt.draugeros.org/aptsync/pool/main/l/linux-upstream kernel
+	rsync -vr rsync://apt.draugeros.org/aptsync/pool/main/l/linux-meta kernel
+	echo -e "\t###\tDELETING CRUFT\t###\t"
+	list=$(ls kernel)
+	for each in $list; do
+		remove=$(ls kernel/$each | grep -v 'amd64.deb$')
+		for each2 in $remove; do
+			rm -rfv kernel/$each/$each2
+		done
 	done
-done
-meta=$(echo kernel/linux-meta/$(ls kernel/linux-meta | sort -Vr | head -1))
-dep=$(dpkg-deb --field $meta Depends | sed 's/, /\\|/g')
-cd kernel/linux-upstream
-rm -rfv $(ls | sed "/\($dep\)/d")
-dep=$(echo "$dep" | sed 's/\\|/ /g' | awk '{print $1}' | sed 's/\(image-\|headers-\)/xanmod_/g')
-rm -rfv $(ls | grep "edge")
-rm -rfv $(ls | grep "cacule")
-cd ../linux-meta
-rm -rfv $(ls | grep -v "$dep")
-rm -rfv $(ls | grep "edge")
-rm -rfv $(ls | grep "cacule")
-cd ..
-dep=$(echo "$dep" | sed 's/xanmod_//g')
-mv linux-upstream "$dep"
-cd ..
-# delete empty folders
-find . -type d -empty -print -delete
-echo -e "\t###\tCOMPRESSING\t###\t"
-tar --verbose --create --xz -f kernel.tar.xz kernel
-echo -e "\t###\tCLEANING\t###\t"
-rm -rfv kernel
-cd ../../../..
+	meta=$(echo kernel/linux-meta/$(ls kernel/linux-meta | sort -Vr | head -1))
+	dep=$(dpkg-deb --field $meta Depends | sed 's/, /\\|/g')
+	cd kernel/linux-upstream
+	rm -rfv $(ls | sed "/\($dep\)/d")
+	dep=$(echo "$dep" | sed 's/\\|/ /g' | awk '{print $1}' | sed 's/\(image-\|headers-\)/xanmod_/g')
+	rm -rfv $(ls | grep "edge")
+	rm -rfv $(ls | grep "cacule")
+	cd ../linux-meta
+	rm -rfv $(ls | grep -v "$dep")
+	rm -rfv $(ls | grep "edge")
+	rm -rfv $(ls | grep "cacule")
+	cd ..
+	dep=$(echo "$dep" | sed 's/xanmod_//g')
+	mv linux-upstream "$dep"
+	cd ..
+	# delete empty folders
+	find . -type d -empty -print -delete
+	echo -e "\t###\tCOMPRESSING\t###\t"
+	tar --verbose --create --xz -f kernel.tar.xz kernel
+	echo -e "\t###\tCLEANING\t###\t"
+	rm -rfv kernel
+	cd ../../..
+fi
+
+# Pshyc - we're compiling shit now
+cd usr/bin
+g++ -fPIE -m64 -o system-installer system-installer.cxx $(python3.9-config --ldflags --cflags --embed)
+cd ../..
 ##############################################################
 #							     #
 #							     #
@@ -90,9 +99,22 @@ if [ -d srv ]; then
 	cp -R srv ../"$FOLDER"/srv
 fi
 cp -R DEBIAN ../"$FOLDER"/DEBIAN
+mkdir -p usr/share/doc/$PAK
+git log > usr/share/doc/$PAK/changelog
+cd usr/share/doc/$PAK
+tar --verbose --create --xz -f changelog.gz changelog 1>/dev/null
+rm changelog
+cd ../../../..
+base="$PWD"
 cd ..
 #DELETE STUFF HERE
-rm system-installer/usr/share/system-installer/modules/kernel.tar.xz
+if [ "$OPTIONS" != "--pool" ]; then
+	rm "$base"/usr/share/system-installer/kernel.tar.xz
+fi
+# delete binary files from repo
+rm "$base"/usr/bin/system-installer
+# delete C++ source from package
+rm "$FOLDER"/usr/bin/system-installer.cxx
 #build the shit
 dpkg-deb --build "$FOLDER"
 rm -rf "$FOLDER"
